backend/requirements.txt
django>=5.2.4
djangorestframework>=3.16.0
gunicorn
